rgy_osdep: Linuxのアフィニティマスク設定が32コア以上を無視するバグを修正
SetProcessAffinityMask/SetThreadAffinityMaskのビット走査で
mask & (1 << j) を使っており、1がint(32bit)のためj>=32で未定義
動作・実質0になり、CPU 32以降が一切設定されなかった。mask_org
復元側は既に(size_t)1u << jだったのと揃え、32論理コア超の
環境(Threadripper/EPYC等)でも正しく動作するよう修正。

+++ NVEncCore/rgy_osdep.h
@@ -332,7 +332,7 @@ static size_t SetProcessAffinityMask(pid_t process, size_t mask) {
     cpu_set_t cpuset;
     CPU_ZERO(&cpuset);
     for (uint32_t j = 0; j < sizeof(mask) * 8; j++) {
-        if (mask & (1 << j)) {
+        if (mask & ((size_t)1u << j)) {
             CPU_SET(j, &cpuset);
         }
     }
@@ -353,7 +353,7 @@ static size_t SetThreadAffinityMask(pthread_t thread, size_t mask) {
     cpu_set_t cpuset;
     CPU_ZERO(&cpuset);
     for (uint32_t j = 0; j < sizeof(mask) * 8; j++) {
-        if (mask & (1 << j)) {
+        if (mask & ((size_t)1u << j)) {
             CPU_SET(j, &cpuset);
         }
     }
